Issue #377 - use the VMUSB drivers in place rather than as copies
Import MDPP with 32 channel QDC firmware support directly from
VMUSB support
