+++ main/mvlc/Makefile.am
@@ -68,11 +68,11 @@ IMPORTED_MODULES=$(IMPORT_DIR)/CMQDC32RdoHdwr.h \
 	$(IMPORT_DIR)/CMADCScaler.h $(IMPORT_DIR)/CMADCScaler.cpp \
 	$(IMPORT_DIR)/CMarker.h $(IMPORT_DIR)/CMarker.cpp \
 	$(IMPORT_DIR)/ CMDPP.h \
-	$(IMPORT_DIR)/CMDPP16QDC.cpp  $(IMPORT_DIR)/CMDPP16QDC.h
+	$(IMPORT_DIR)/CMDPP16QDC.cpp  $(IMPORT_DIR)/CMDPP16QDC.h \
+	$(IMPORT_DIR)/CMDPP32QDC.cpp  	$(IMPORT_DIR)/CMDPP32QDC.h 
 
 LOCAL_MODULES= 	MADC32Registers.h \
 	CMTDC32.h CMTDC32.cpp \
-	CMDPP32QDC.cpp  CMDPP32QDC.h \
 	CMDPP32SCP.cpp  CMDPP32SCP.h   CMDPPQDC.h \
 	CNADC2530.h CNADC2530.cpp CV977.h CV977.cpp \
 	CV1495sc.h CV1495sc.cpp CV1729.h CV1729.cpp \
