+++ main/usb/vmusb/daqconfig/CMDPP32QDC.cpp
@@ -14,12 +14,17 @@
 	     East Lansing, MI 48824-1321
 */
 
-#include "CMDPP32QDC.h"
-#include "CReadoutModule.h"
+#include <CMDPP32QDC.h>
+#include <CReadoutModule.h>
 #include <unistd.h>
 #include <CVMUSB.h>
 #include <bitset>
 #include <iomanip>
+#include <iostream>
+
+#ifdef MVLC_GENERATOR
+#include <XXUSBConfigurableObject.h>
+#endif
 
 using std::vector;
 using std::string;
@@ -59,7 +64,7 @@ CMDPP32QDC::CMDPP32QDC()
 {
   m_pConfiguration = 0;
 }
-
+#ifndef MVLC_GENERATOR
 /**
  * Copy construction.  This cannot be virtual by the rules of C++ the clone()
  * method normally creates a new object from an existing template object.
@@ -73,6 +78,7 @@ CMDPP32QDC::CMDPP32QDC(const CMDPP32QDC& rhs)
     m_pConfiguration = new CReadoutModule(*(rhs.m_pConfiguration));
   }
 }
+#endif
 /**
  * Destruction.  If your object creatd any dynamic data it must be freed here:
  */
@@ -97,7 +103,11 @@ CMDPP32QDC::~CMDPP32QDC()
  * @parm configuration - Reference to the configuration object for this instance of the driver.
  */
 void
+#ifdef MVLC_GENERATOR
+CMDPP32QDC::onAttach(XXUSB::CConfigurableObject& configuration)
+#else
 CMDPP32QDC::onAttach(CReadoutModule& configuration)
+#endif
 {
   m_pConfiguration = &configuration; 
 
@@ -167,17 +177,24 @@ CMDPP32QDC::Initialize(CVMUSB& controller)
 
   // Retreiving trigger information before the module reset
   uint16_t triggersource = m_pConfiguration -> getIntegerParameter("-triggersource");
+#ifndef MVLC_GENERATOR
   if (triggersource == 0x400) {
     controller.vmeRead16(base + TriggerSource, initamod, &triggersource);
   }
-
+#endif
   uint16_t triggeroutput = m_pConfiguration -> getIntegerParameter("-triggeroutput");
+#ifndef MVLC_GENERATOR
   if (triggeroutput == 0x400) {
     controller.vmeRead16(base + TriggerOutput, initamod, &triggeroutput);
   }
+#endif
 
   controller.vmeWrite16(base + Reset,        initamod, 1);
+#ifdef MVLC_GENERATOR
+  controller.delay(1000);
+#else
   sleep(1);
+#endif
   controller.vmeWrite16(base + StartAcq,     initamod, 0);
   controller.vmeWrite16(base + InitFifo,     initamod, 1);
   controller.vmeWrite16(base + ReadoutReset, initamod, 1);
@@ -325,13 +342,16 @@ CMDPP32QDC::Initialize(CVMUSB& controller)
   char readBuffer[100];		// really a dummy as these are all write...
   size_t bytesRead;
   int status = controller.executeList(list, readBuffer, sizeof(readBuffer), &bytesRead);
+#ifndef MVLC_GENERATOR
   if (status < 0) {
      throw string("List excecution to initialize an MDPP32QDC failed");
   }
-
+#endif
+#ifndef MVLC_GENERATOR
   if (isPrintRegisters) {
     printRegisters(controller);
   }
+#endif
 }
 
 /**
@@ -364,7 +384,7 @@ void
 CMDPP32QDC::onEndRun(CVMUSB& controller)
 {
 }
-
+#ifndef MVLC_GENERATOR
 /**
  * This method virtualizes copy construction by providing a virtual method that
  * invokes it. Usually you don't have to modify this code.
@@ -378,6 +398,7 @@ CMDPP32QDC::clone() const
 {
   return new CMDPP32QDC(*this);
 }
+#endif
 
 /*
   Creates a map from the value of -gaincorrectionlong and -gaincorrectionshort
@@ -507,7 +528,7 @@ CMDPP32QDC::initCBLTReadout(CVMUSB& controller,
   controller.vmeWrite16(cbltAddress + ReadoutReset, initamod, 0);
   controller.vmeWrite16(cbltAddress + StartAcq,     initamod, 1);
 }
-
+#ifndef MVLC_GENERATOR
 /**
  * Printing all register values in MDPP-32 module with QDC firmware
  * read from the module, not the user-input values.
@@ -876,4 +897,33 @@ CMDPP32QDC::printRegisters(CVMUSB& controller)
     cout << endl;
   }
 }
+#endif
 //////////////////////////////////////////////////////////////////////////////////////////
+#ifdef MVLC_GENERATOR
+////////////////////////////// Mdpp32qdcCommand implementation
+
+
+/**
+ *  constructor 
+ */
+Mdpp32qdcCommand::Mdpp32qdcCommand(CTCLInterpreter& interp, TCLConfigParser& parser) :
+  DeviceCommand(interp, "mdpp32qdc", parser) {}
+
+/**
+ * destructor
+ */
+Mdpp32qdcCommand::~Mdpp32qdcCommand() {}
+
+/** createDevice
+ * 
+ * 
+ */
+CReadoutModule*
+Mdpp32qdcCommand::createDevice(std::string name) {
+  CReadoutModule* result = new CReadoutModule;
+  result->SetDriver(new CMDPP32QDC);
+
+  return result;
+}
+
+#endif
